requirements.txt
discord
requests
beautifulsoup4
python-dotenv